Distributed Tracing and logging
Zipkin is a distributed tracing system. It helps gather timing data needed to troubleshoot latency problems in service architectures. It is a visualization of your request-response and trace it. ELK stack is used for distributed logging. The Zipkin server gives a visual aid of your request-response and timing details, while Kibana in ELK gives a visual aid for centralized logging - somewhere where you can read the logs